change timer to emit double values, added test to play with calculations

+++ app/src/main/java/name/mharbovskyi/satellitegame/domain/ObjectStateTransformer.kt
@@ -4,5 +4,5 @@ import name.mharbovskyi.satellitegame.domain.entity.ObjectState
 import name.mharbovskyi.satellitegame.domain.entity.Planet
 import name.mharbovskyi.satellitegame.domain.physics.nextObjectState
 
-fun objectStateTransformer(planet: Planet): (ObjectState, Long) -> ObjectState =
+fun objectStateTransformer(planet: Planet): (ObjectState, Double) -> ObjectState =
     {satellite, timeInterval -> nextObjectState(satellite, planet, timeInterval)}
